+++ packages/wallets/src/evm/index.ts
@@ -1,2 +1,4 @@
 export * from "./wallet";
 export * from "./chains";
+
+export type { EVMSmartWalletChain } from "./chains";