+++ packages/wallets/src/index.ts
@@ -1,1 +1,7 @@
 export * from "./sdk";
+export { SolanaSmartWallet, SolanaMPCWallet } from "./solana/wallet";
+export { EVMSmartWallet, EVMMPCWallet } from "./evm/wallet";
+
+export type { EVMSignerInput, EVMSigner, EVMSmartWalletChain } from "./evm";
+export type { SolanaSignerInput } from "./solana";
+export type { WalletTypeToArgs } from "./services/types";