+++ packages/wallets/src/solana/index.ts
@@ -1,1 +1,3 @@
 export * from "./wallet";
+export * from "./tokens";
+export * from "./types/signers";