+++ packages/wallets/src/utils/constants.ts
@@ -3,3 +3,5 @@ import { name, version } from "../../package.json";
 export const WALLETS_SERVICE = "WALLETS_SDK";
 export const SDK_VERSION = version;
 export const SDK_NAME = name;
+export const ENTRY_POINT_ADDRESS = "0x0000000071727de22e5e9d8baf0edac6f37da032";
+export const STATUS_POLLING_INTERVAL_MS = 1000;