+++ apps/auth/nextjs-ssr/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @crossmint/server-sdk-next-starter
 
+## 0.4.25
+
+### Patch Changes
+
+- @crossmint/client-sdk-react-ui@1.16.1
+
+## 0.4.24
+
+### Patch Changes
+
+- Updated dependencies [025a5a2]
+  - @crossmint/client-sdk-react-ui@1.16.0
+
+## 0.4.23
+
+### Patch Changes
+
+- Updated dependencies [063ae65]
+  - @crossmint/client-sdk-react-ui@1.15.3
+  - @crossmint/server-sdk@1.2.9
+
 ## 0.4.22
 
 ### Patch Changes