+++ packages/client/ui/react-ui/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @crossmint/client-sdk-react-ui
 
+## 1.16.1
+
+### Patch Changes
+
+- Updated dependencies [ca39114]
+  - @crossmint/wallets-sdk@0.2.2
+
+## 1.16.0
+
+### Minor Changes
+
+- 025a5a2: Migrated to new wallets-sdk package for evm and solana smart wallets.
+
+### Patch Changes
+
+- Updated dependencies [9c71ce8]
+  - @crossmint/wallets-sdk@0.2.1
+
+## 1.15.3
+
+### Patch Changes
+
+- 063ae65: Upgrade dynamic to v4
+- Updated dependencies [063ae65]
+  - @crossmint/client-sdk-smart-wallet@0.3.1
+  - @crossmint/client-sdk-base@1.4.12
+  - @crossmint/client-sdk-auth@1.1.21
+  - @crossmint/common-sdk-auth@1.0.19
+
 ## 1.15.2
 
 ### Patch Changes